Button's logic implement

+++ Calculadora/Calcs.cs
@@ -8,25 +8,25 @@ namespace Calculadora
 {
     class Calcs
     {
-        public int Sum(int n1, int n2)
+        public double Sum(double n1, double n2)
         {
-            int n3;
+            double n3;
             n3 = n1 + n2;
 
             return n3;
         }
 
-        public int Sub(int n1, int n2)
+        public double Sub(double n1, double n2)
         {
-            int n3;
+            double n3;
             n3 = n1 - n2;
 
             return n3;
         }
 
-        public int Mul(int n1, int n2)
+        public double Mul(double n1, double n2)
         {
-            int n3;
+            double n3;
             n3 = n1 * n2;
 
             return n3;
